requirements.txt
pandas
aiohttp
asyncio
bs4
dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Brotli